+++ literary_system/llm_bridge/__init__.py
@@ -1,0 +1,14 @@
+
+from literary_system.llm_bridge.canonical_bridge_v2 import (
+    BridgeConfig,
+    BridgeResponse,
+    CanonicalBridgeV2,
+    ModelType,
+)
+
+__all__ = [
+    "BridgeConfig",
+    "BridgeResponse",
+    "CanonicalBridgeV2",
+    "ModelType",
+]